+++ CHANGELOG.md
@@ -34,6 +34,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Fixed
 
+- Manage certificate linked to an enrollment
 - Certificate products are not listed anymore in teacher dashboard
 - Fix several SaleTunnel cache issues
 - Fix course access link in learner dashboard. This link must not be